# src/AterStudio/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ater Studio",
    "description": "API 文档",
    "version": "v1"
  },
  "paths": {
...
    "/api/Entity/{id}": {
      "get": {
        "tags": [
          "Entity"
        ],
        "operationId": "Entity_List",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "serviceName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityFile"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityFile"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityFile"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Entity/dtos": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "s\r\n            获取dtos",
        "operationId": "Entity_GetDtos",
        "parameters": [
          {
            "name": "entityFilePath",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityFile"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityFile"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityFile"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Entity/dto": {
      "post": {
        "tags": [
          "Entity"
        ],
        "summary": "创建DTO",
        "operationId": "Entity_CreateDto",
        "parameters": [
          {
            "name": "entityFilePath",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "summary",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entity"
        ],
        "summary": "更新内容",
        "operationId": "Entity_UpdateDtoContent",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDtoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDtoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDtoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Entity/fileContent": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "获取文件内容 entity/manager",
        "operationId": "Entity_GetFileContent",
        "parameters": [
          {
            "name": "entityName",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isManager",
            "in": "query",
            "description": "是否为manager",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "moduleName",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityFile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityFile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityFile"
                }
              }
            }
          }
        }
      }
    },
    "/api/Entity/generate": {
      "post": {
        "tags": [
          "Entity"
        ],
        "operationId": "Entity_Generate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Entity/batch-generate": {
      "post": {
        "tags": [
          "Entity"
        ],
        "summary": "批量生成",
        "operationId": "Entity_BatchGenerate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchGenerateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchGenerateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BatchGenerateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Entity/generateRequest/{id}": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "生成前端请求",
        "operationId": "Entity_GenerateRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "webPath",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "schema": {
              "$ref": "#/components/schemas/RequestLibType"
            }
          },
          {
            "name": "swaggerPath",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Entity/generateClientRequest/{id}": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "生成客户端请求",
        "operationId": "Entity_GenerateClientRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "webPath",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "schema": {
              "$ref": "#/components/schemas/LanguageType"
            }
          },
          {
            "name": "swaggerPath",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Entity/generateSync/{id}": {
      "post": {
        "tags": [
          "Entity"
        ],
        "summary": "同步ng页面",
        "operationId": "Entity_GenerateSync",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Entity/generateNgModule/{id}": {
      "post": {
        "tags": [
          "Entity"
        ],
        "summary": "生成NG组件模块",
        "operationId": "Entity_GenerateNgModule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entityName",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rootPath",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchGenerateDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "entityPaths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "serviceName": {
            "type": "string",
            "description": "服务",
            "nullable": true
          },
          "commandType": {
            "$ref": "#/components/schemas/CommandType"
          },
          "projectPath": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "项目路径",
            "nullable": true
          },
          "force": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CommandType": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "description": "命令类型",
        "format": "int32",
        "x-enumNames": [
          "Dto",
          "Manager",
          "API",
          "Protobuf",
          "Clear"
        ],
        "x-enumData": [ ]
      },
...
      "EntityFile": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "comment": {
            "type": "string",
            "description": "注释说明",
            "nullable": true
          },
          "baseDirPath": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "module": {
            "type": "string",
            "description": "所属模块",
            "nullable": true
          },
          "hasDto": {
            "type": "boolean"
          },
          "hasManager": {
            "type": "boolean"
          },
          "hasAPI": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "GenerateDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "entityPath": {
            "type": "string"
          },
          "serviceName": {
            "type": "string",
            "description": "服务",
            "nullable": true
          },
          "commandType": {
            "$ref": "#/components/schemas/CommandType"
          },
          "force": {
            "type": "boolean",
            "description": "是否覆盖"
          }
        },
        "additionalProperties": false
      },
      "LanguageType": {
        "enum": [
          0
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "CSharp"
        ],
        "x-enumData": [ ]
      },
...
      "RequestLibType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "NgHttp",
          "Axios"
        ],
        "x-enumData": [ ]
      },
...
      "UpdateDtoDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "maxLength": 100,
            "type": "string",
            "description": "文件完整路径"
          },
          "content": {
            "maxLength": 2000,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}